unitree/requirements.txt
api==0.0.7
asgiref==3.6.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2.1
django-ckeditor==6.5.1
django-embed-video==1.4.8
django-filter==23.2
django-heroku==0.3.1
django-js-asset==2.0.0
djangorestframework==3.14.0
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
nose==1.3.7
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
whitenoise==6.4.0
